refactor(agents-desktop): extract main process utilities
Move low-coupling header, server normalization, and runtime health helpers out of main.ts so later lifecycle extraction has smaller dependencies.
